split macos startup to different executable
